experimental: use a map and frames to keep track of diffs
This is broken because it throws out the only known full states when using
diffs to compute far states. In this model, you always need to keep the
latest states around in order to re-construct any state.

This isn't ideal because the only *full* known states are the latest two.
If you've already done rewind, and if you want to rewind again (to some further
place), you need to compute everything all over again.

Consider the following options:

	1. Always keep *only* the full initial state [frame -123], then always
	   take diffs from the initial state. Reconstruct states by using the
	   initial state and succesively applying diffs.

	2. Always keep a full initial state [frame -123], *and* a latest state.
	   If you need to rewind very far, apply diffs to the initial state
	   instead of the latest state.